refactor(act): rename adapter files to kebab-case
The rest of the codebase uses kebab-case for filenames. The three
adapter files (ConsoleLogger.ts, InMemoryCache.ts, InMemoryStore.ts)
were the only PascalCase outliers. Rename to match and update every
import path in the source and tests.

Pure rename — no behavior change.
