+++ libs/act/src/adapters/console-logger.ts
