+++ libs/act/test/tracing.spec.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
-import { InMemoryCache } from "../src/adapters/InMemoryCache.js";
-import { InMemoryStore } from "../src/adapters/InMemoryStore.js";
+import { InMemoryCache } from "../src/adapters/in-memory-cache.js";
+import { InMemoryStore } from "../src/adapters/in-memory-store.js";
 import { state } from "../src/builders/state-builder.js";
 import * as drain from "../src/internal/drain.js";
 import * as es from "../src/internal/event-sourcing.js";
